Code refactoring

+++ philo/ft_utilities1.c
@@ -6,7 +6,7 @@
 /*   By: fabet <fabet@student.21-school.ru>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/01 12:15:49 by fabet             #+#    #+#             */
-/*   Updated: 2022/07/01 21:31:26 by fabet            ###   ########.fr       */
+/*   Updated: 2022/07/01 23:24:19 by fabet            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,12 +33,12 @@ int	ft_strict_atoi(const char *str)
 	return (result);
 }
 
-long	ft_count_timestamp_in_ms(
+int	ft_count_timestamp_in_ms(
 			t_timeval start_time,
 			t_timeval actual_time)
 {
-	long	seconds;
-	long	useconds;
+	int	seconds;
+	int	useconds;
 
 	seconds = actual_time.tv_sec - start_time.tv_sec;
 	useconds = actual_time.tv_usec - start_time.tv_usec;
